+++ main-server/src/http_server/routes/v4/rp.js
@@ -73,7 +73,8 @@ router.post(
         },
         {
           synchronous: false,
-          apiVersion: apiVersion,
+          validateIdentifier: false,
+          apiVersion,
         }
       );